Complete Task 3.2: Set up Koin DI for iOS
Subtask 3.2.1: Initialize Koin in iOS app
- Created DIHelper.kt with initKoin() and doInitKoin() functions
- Updated CocktailCraftApp.swift to initialize Koin on startup
- Configured Koin modules with appModule + platformModule()

Subtask 3.2.3: Implement ViewModel access from SwiftUI
- Created KoinHelper.kt to provide iOS access to ViewModels
- Implemented ObservableViewModel base class for KMP ViewModel wrapping
- Created FlowCollector to bridge Kotlin StateFlow to Swift
- Implemented specific Observable wrappers:
  - ObservableHomeViewModel with cocktail list functionality
  - ObservableCocktailDetailViewModel with detail operations
  - ObservableCartViewModel with cart management
  - ObservableThemeViewModel with theme toggling
- Created ViewModelProvider for centralized ViewModel access
- Updated ContentView to demonstrate:
  - Loading cocktails from shared HomeViewModel
  - Displaying loading states and error messages
  - Theme toggle functionality
  - Proper SwiftUI integration with reactive updates

Also fixed iOS platform implementations:
- NetworkMonitor: Simplified implementation using URLSession
- ImageLoader: Fixed ULong type conversion for NSURLCache

The iOS app can now access and use shared ViewModels with proper
reactive updates in SwiftUI.

+++ shared/src/iosMain/kotlin/com/cocktailcraft/util/ImageLoader.kt
@@ -17,8 +17,8 @@ actual class ImageLoader {
         val config = NSURLSessionConfiguration.defaultSessionConfiguration
         // Configure cache: 100MB disk cache, 10MB memory cache
         config.URLCache = NSURLCache(
-            memoryCapacity = 10u * 1024u * 1024u, // 10 MB
-            diskCapacity = 100u * 1024u * 1024u,   // 100 MB
+            memoryCapacity = (10 * 1024 * 1024).toULong(), // 10 MB
+            diskCapacity = (100 * 1024 * 1024).toULong(),   // 100 MB
             diskPath = "cocktailcraft_image_cache"
         )
         config.requestCachePolicy = NSURLRequestReturnCacheDataElseLoad
